pixel-lab/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-multipart>=0.0.9
httpx>=0.27
Pillow>=10.3
numpy>=1.26
opencv-python-headless>=4.9
scipy>=1.12
